B1946 신입사원 보류

+++ baekjoon/src/b13305_주유소/Main.java
@@ -6,8 +6,8 @@ import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
 class Node{
-	int price;
-	int length;
+	long price;
+	long length;
 	
 	
 }
@@ -18,7 +18,7 @@ public class Main {
 		
 		int N = Integer.parseInt(br.readLine());
 		Node[] node = new Node[N]; 
-		int sum = 0 ;
+		long sum = 0;
 		
 		// 객체 배열 초기화.
 		for(int i = 0 ; i < N; i++) {
@@ -36,11 +36,10 @@ public class Main {
 		}
 		
 		sum = node[0].price*node[0].length;
-		int minPrice = node[0].price ;
+		long minPrice = node[0].price ;
 		for(int i = 1 ; i < node.length; i++) {
 			if(node[i].price < node[i-1].price ) {
 				minPrice = Math.min(minPrice , node[i].price);
-				
 			}
 			sum += (minPrice * node[i].length);
 		}
